+++ internal/controller/user/user_subscription_subscription_update.go
@@ -21,7 +21,7 @@ func (c *ControllerSubscription) Update(ctx context.Context, req *subscription.U
 	if !config.GetConfigInstance().IsLocal() {
 		//User 检查
 		utility.Assert(_interface.Context().Get(ctx).User != nil, "auth failure,not login")
-		utility.Assert(int64(_interface.Context().Get(ctx).User.Id) == sub.UserId, "userId not match")
+		utility.Assert(_interface.Context().Get(ctx).User.Id == sub.UserId, "userId not match")
 	}
 	return service.SubscriptionUpdate(ctx, req, 0)
 }
