+++ internal/controller/user/user_subscription_subscription_update_preview.go
@@ -22,7 +22,7 @@ func (c *ControllerSubscription) UpdatePreview(ctx context.Context, req *subscri
 	if !config.GetConfigInstance().IsLocal() {
 		//User 检查
 		utility.Assert(_interface.Context().Get(ctx).User != nil, "auth failure,not login")
-		utility.Assert(int64(_interface.Context().Get(ctx).User.Id) == sub.UserId, "userId not match")
+		utility.Assert(_interface.Context().Get(ctx).User.Id == sub.UserId, "userId not match")
 	}
 	prepare, err := service.SubscriptionUpdatePreview(ctx, req, 0, 0)
 	if err != nil {
